requirements.txt
Flask
requests
pycryptodome
protobuf